last call

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -9,7 +9,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """return the list of all the delays in ascending order"""
     list_of_delays: List[float] = []
     for _ in range(n):
